added vlans feature
Adds the ability to manage VLAN ids. Pools of 4096 VLANs can be created and
sub-divided into Partitions. A VLAN can then be requested by a user from a Partition, and
they will receive an unused VLAN id, which will then be marked as used until they
release it. Projects can be assigned Partition's and a system-wide default Partition can
be set for users to request VLAN Ids from their Project's Partition or the default as well.

- VLANs can have tags for organizational purposes and can be requsted by tag
- A VLAN can be requested by specific vlanId within a Partition
- VLANs can be marked as reserved (0, 1, and 4095 are reserved by default) so that they
will never be used
- fixed modified properties in entity updated events to restore signalR functionality

Co-authored-by: Nicholas Artie John Clark <naclark@sei.cmu.edu>
Co-authored-by: Andrew Schlackman <72105194+sei-aschlackman@users.noreply.github.com>