added azure destroy failure handling
Attempts to mitigate common issues with destroying Azure workspaces. If previous
destroys failed, remove from the state any resources that exist within an
azurerm_resource_group managed by this workspace. They will be destroyed implicitly
when destroying the workspace and can avoid errors with destroying those resources
directly.

This behavior can be disabled or the number of failures before taking action
adjusted by changing new properties on Directory and Workspace entities. When a
Workspace is created, it's Threshold value will be set to that of it's first Ancestor Directory
that has a value set. If no values are found, the appsetting will be used. If an Ancestor
Directory has AzureDestroyFailureThresholdEnabled set to false before a value has been
found, the Threshold will be set to null on the Workspace and this new behavior will
not trigger. The value set on the Workspace is always used for Runs in that Workspace,
regardless of changes made to Directories or appsettings after the Workspace has
been created.

- added AzureDestroyFailureThreshold to Directory and Workspace
- added AzureDestroyFailureThreshhold to appsettings
- added AzureDestroyFailureThresholdEnabled to Directory
- fixed getting details of Azure Resource by url decoding the Id